refactor: keyword type 공통으로 변경
