+++ src/app/create/types/keywordChipType.ts
@@ -1,0 +1,1 @@
+export type KeywordChipType = "default" | "add";
